+++ packages/frontend/web/src/components/ThemeSwitcher.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 
-import DarkIcone from './icons/DarkIcone';
-import LightIcon from './icons/LightIcone';
+import DarkIcon from './icons/DarkIcon';
+import LightIcon from './icons/LightIcon';
 
 export default function ThemeSwitcher() {
   const [isDarkMode, setIsDarkMode] = useState(false);
@@ -16,12 +16,12 @@ export default function ThemeSwitcher() {
 
   return (
     <div
-      className='dark:bg-dark-ulta-light bg-light-light active:shadow-secondary dark:active:shadow-dark-medium flex h-8 w-8 items-center justify-center rounded-full shadow-md duration-200 active:translate-y-[2px] active:shadow-inner md:h-10 md:w-10'
+      className='bg-light-light active:shadow-divider-dark flex h-8 w-8 items-center justify-center rounded-full shadow-md duration-200 active:translate-y-[2px] active:shadow-inner md:h-10 md:w-10'
       onClick={toggleDarkMode}
     >
-      <DarkIcone className='text-light-light h-4 w-4 md:h-5 md:w-5' />
+      <DarkIcon className='text-secondary scale-moon h-4 w-4 md:h-5 md:w-5' />
 
-      <LightIcon className='text-secondary absolute h-4 w-4 dark:text-transparent md:h-5 md:w-5' />
+      <LightIcon className='text-secondary scale-sun absolute h-4 w-4 md:h-5 md:w-5' />
     </div>
   );
 }
