+++ packages/frontend/web/src/components/icons/LightIcon.tsx
