+++ packages/frontend/web/src/globals.css
@@ -4,23 +4,34 @@
 @tailwind components;
 @tailwind utilities;
 
+@layer base {
+  :root {
+    --primary: #eb0573;
+    --secondary: #3f436a;
+    --light-hard: #eae3e3;
+    --light-medium: #f4efef;
+    --light-light: #fcfcfc;
+    --moon: 0;
+    --sun: 1;
+  }
+
+  .dark {
+    --primary: #bd0069;
+    --secondary: #fcfcfe;
+    --light-hard: #111214;
+    --light-medium: #1c1c1f;
+    --light-light: #272721;
+    --moon: 1;
+    --sun: 0;
+  }
+
+  body {
+    @apply bg-light-medium;
+  }
+}
+
 @layer utilities {
   .logo-size {
     width: clamp(24px, 4vw, 40px);
   }
-}
-
-body {
-  height: 100dvh;
-  width: 100dvw;
-  margin: 0;
-  padding: 0;
-}
-
-.dark {
-  background-color: #1c1c1f;
-}
-
-.light {
-  background-color: #f4efef;
 }
