+++ packages/frontend/web/tailwind.config.js
@@ -5,18 +5,13 @@ export default {
   theme: {
     extend: {
       colors: {
-        primary: '#eb0573',
-        'primary-dark': '#bd0069',
-        secondary: '#3f436a',
-        'dark-hard': '#111214',
-        'dark-medium': '#1c1c1f',
-        'dark-light': '#272721',
+        primary: 'var(--primary)',
+        secondary: 'var(--secondary)',
+        'light-hard': 'var(--light-hard)',
+        'light-medium': 'var(--light-medium)',
+        'light-light': 'var(--light-light)',
         'dark-ulta-light': '#3c4045',
-        'light-hard': '#eae3e3',
-        'light-medium': '#f4efef',
-        'light-light': '#fcfcfc',
         'divider-dark': '#4d4f58',
-        'divider-light': '#ccced7',
         placeholder: '#9c89a2',
         success: '#51d686',
         'super-like': '#59c3ff',
@@ -25,6 +20,10 @@ export default {
       fontFamily: {
         title: ['Heavitas'],
         base: ['Poppins'],
+      },
+      scale: {
+        moon: 'var(--moon)',
+        sun: 'var(--sun)',
       },
     },
   },
